pi/requirements.txt
discord.py
wakeonlan
paramiko
